requirements.txt
google-adk
fastapi
uvicorn
python-dotenv
google-genai
psycopg2-binary